requirements.txt
Django
asgiref
google-generativeai
whitenoise
gunicorn
Pillow
python-dotenv
requests